requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==23.2.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
asttokens==3.0.0
attrs==24.3.0
basicsr==1.4.2
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
einops==0.7.0
executing==2.1.0
facexlib==0.3.0
fastapi==0.115.6
ffmpy==0.4.0
filelock==3.13.1
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.55.3
fsspec==2024.2.0
future==1.0.0
gfpgan==1.3.8
glcontext==2.5.0
gradio==5.9.1
gradio_client==1.5.2
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.5.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
moderngl==5.10.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
onnxruntime-gpu==1.20.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.48
protobuf==5.29.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
PyMatting==1.1.13
pymeshlab==2023.12.post2
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
realesrgan==0.3.0
referencing==0.35.1
regex==2024.11.6
rembg==2.0.60
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
ruff==0.8.3
safehttpx==0.1.6
safetensors==0.4.5
scikit-image==0.25.0
scipy==1.14.1
semantic-version==2.10.0
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
super-image==0.1.7
sympy==1.13.1
tb-nightly==2.19.0a20241217
tensorboard-data-server==0.7.2
tifffile==2024.12.12
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchmcubes @ git+https://github.com/tatsy/torchmcubes.git@cb81cddece46a8a126b08f7fbb9742f8605eefab
torchvision==0.20.1+cu118
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
trimesh==4.0.5
triton==3.1.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.1.3
wheel==0.45.1
wxPython==4.2.2
xatlas==0.0.9
yapf==0.43.0
